12 batch update mappings (#16)
* linter

* Added batch mappings update.

We have now a PUT /mappings accepting a dict with remove and add entries to
batch remove and add mappings.
This will be used from the update_gen script in the APs.

* New JSON format with regex

* WIP Refactor permissions usage

* Fix remove permissions into syncpermissions

* Fix bug with the add-remove order in two calls.

In syncing local permissions and the manager, we added and then removed targets and that is wrong. We should remove and only then add new ones.