Loading configuration from the binary directory.
When loading the YML configuration file, refer to its path as seen from
the directory of the prometheus binary. We essentially need to know the
common prefix between both directories (config and binary files) and to
chdir to the bin directory beforehand.
Encapsulated ConfigManager to recompute the common prefix only when the
variables change.